+++ src/main/java/org/sonarsource/plugins/bml/measures/SetSizeOnFilesSensor.java
@@ -17,26 +17,37 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.plugins.example.settings;
+package org.sonarsource.plugins.bml.measures;
 
+import static org.sonarsource.plugins.bml.measures.ExampleMetrics.FILENAME_SIZE;
+
+import org.sonar.api.batch.fs.FileSystem;
+import org.sonar.api.batch.fs.InputFile;
 import org.sonar.api.batch.sensor.Sensor;
 import org.sonar.api.batch.sensor.SensorContext;
 import org.sonar.api.batch.sensor.SensorDescriptor;
-import org.sonar.api.utils.log.Loggers;
 
-public class SayHelloFromScanner implements Sensor {
-
+/**
+ * Scanner feeds raw measures on files but must not aggregate values to directories and project.
+ * This class emulates loading of file measures from a 3rd-party analyser.
+ */
+public class SetSizeOnFilesSensor implements Sensor {
   @Override
   public void describe(SensorDescriptor descriptor) {
-    descriptor.name(getClass().getName());
+    descriptor.name("Compute size of file names");
   }
 
   @Override
   public void execute(SensorContext context) {
-    if (context.config().getBoolean(HelloWorldProperties.HELLO_KEY).orElse(false)) {
-      // print log only if property is set to true
-      Loggers.get(getClass()).info("Hello World!");
+    FileSystem fs = context.fileSystem();
+    // only "main" files, but not "tests"
+    Iterable<InputFile> files = fs.inputFiles(fs.predicates().hasType(InputFile.Type.MAIN));
+    for (InputFile file : files) {
+      context.<Integer>newMeasure()
+        .forMetric(FILENAME_SIZE)
+        .on(file)
+        .withValue(file.filename().length())
+        .save();
     }
   }
-
 }
