+++ src/main/java/org/sonarsource/plugins/bml/rules/FlagLineSensor.java
@@ -1,4 +1,4 @@
-package org.sonarsource.plugins.example.rules;
+package org.sonarsource.plugins.bml.rules;
 
 import org.sonar.api.batch.fs.FilePredicates;
 import org.sonar.api.batch.fs.InputFile;
@@ -7,7 +7,7 @@ import org.sonar.api.batch.rule.Checks;
 import org.sonar.api.batch.sensor.Sensor;
 import org.sonar.api.batch.sensor.SensorContext;
 import org.sonar.api.batch.sensor.SensorDescriptor;
-import org.sonarsource.plugins.example.languages.FooLanguage;
+import org.sonarsource.plugins.bml.languages.BMLLanguage;
 
 public class FlagLineSensor implements Sensor {
 
@@ -15,20 +15,20 @@ public class FlagLineSensor implements Sensor {
 
   public FlagLineSensor(CheckFactory checkFactory) {
     checks = checkFactory.create(FlagRuleDefinition.REPO_KEY);
-    checks.addAnnotatedChecks(FlagLine1Rule.class, FlagLine2Rule.class, FlagLine3Rule.class);
+    checks.addAnnotatedChecks(CommentChecker.class,MultipleForLoop.class);
   }
 
   @Override
   public void describe(SensorDescriptor descriptor) {
-    descriptor.name(FlagLine1Rule.RULE_KEY + "sensor");
-    descriptor.onlyOnLanguages(FooLanguage.KEY);
+    descriptor.name(CommentChecker.RULE_KEY + "sensor");
+    descriptor.onlyOnLanguages(BMLLanguage.KEY);
     descriptor.createIssuesForRuleRepository(FlagRuleDefinition.REPO_KEY);
   }
 
   @Override
   public void execute(SensorContext context) {
     FilePredicates p = context.fileSystem().predicates();
-    for (InputFile inputFile : context.fileSystem().inputFiles(p.hasLanguages(FooLanguage.KEY))) {
+    for (InputFile inputFile : context.fileSystem().inputFiles(p.hasLanguages(BMLLanguage.KEY))) {
       checks.all().forEach(check -> check.execute(context, inputFile, checks.ruleKey(check)));
     }
   }
