+++ src/main/java/org/sonarsource/plugins/bml/rules/MultipleForLoop.java
@@ -1,4 +1,4 @@
-package org.sonarsource.plugins.example.rules;
+package org.sonarsource.plugins.bml.rules;
 
 import org.sonar.api.batch.fs.InputFile;
 import org.sonar.api.batch.sensor.SensorContext;
@@ -6,18 +6,18 @@ import org.sonar.api.batch.sensor.issue.NewIssue;
 import org.sonar.api.rule.RuleKey;
 import org.sonar.check.Rule;
 
-@Rule(key = FlagLine2Rule.RULE_KEY, name = "Example rule 2", description = "Example rule 2 description")
-public class FlagLine2Rule implements FlagLineRule {
-  public static final String RULE_KEY = "ExampleRule2";
+@Rule(key = MultipleForLoop.RULE_KEY, name = "Multiple For Loop ", description = "File shall not contain MultipleForLoop")
+public class MultipleForLoop implements FlagLineRule {
+  public static final String RULE_KEY = "MultipleForLoop";
 
   @Override
   public void execute(SensorContext sensorContext, InputFile file, RuleKey ruleKey) {
-    NewIssue newIssue = sensorContext.newIssue();
+	NewIssue newIssue = sensorContext.newIssue();
     newIssue
       .forRule(ruleKey)
       .at(newIssue.newLocation()
         .on(file)
-        .at(file.selectLine(2)))
+        .at(file.selectLine(1)))
       .save();
   }
 }
