+++ src/main/java/org/sonarsource/plugins/bml/settings/HelloWorldProperties.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.plugins.example.settings;
+package org.sonarsource.plugins.bml.settings;
 
 import java.util.List;
 import org.sonar.api.config.PropertyDefinition;
