+++ src/main/java/org/sonarsource/plugins/bml/settings/SayHelloFromScanner.java
@@ -17,37 +17,26 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.plugins.example.measures;
+package org.sonarsource.plugins.bml.settings;
 
-import org.sonar.api.batch.fs.FileSystem;
-import org.sonar.api.batch.fs.InputFile;
 import org.sonar.api.batch.sensor.Sensor;
 import org.sonar.api.batch.sensor.SensorContext;
 import org.sonar.api.batch.sensor.SensorDescriptor;
+import org.sonar.api.utils.log.Loggers;
 
-import static org.sonarsource.plugins.example.measures.ExampleMetrics.FILENAME_SIZE;
+public class SayHelloFromScanner implements Sensor {
 
-/**
- * Scanner feeds raw measures on files but must not aggregate values to directories and project.
- * This class emulates loading of file measures from a 3rd-party analyser.
- */
-public class SetSizeOnFilesSensor implements Sensor {
   @Override
   public void describe(SensorDescriptor descriptor) {
-    descriptor.name("Compute size of file names");
+    descriptor.name(getClass().getName());
   }
 
   @Override
   public void execute(SensorContext context) {
-    FileSystem fs = context.fileSystem();
-    // only "main" files, but not "tests"
-    Iterable<InputFile> files = fs.inputFiles(fs.predicates().hasType(InputFile.Type.MAIN));
-    for (InputFile file : files) {
-      context.<Integer>newMeasure()
-        .forMetric(FILENAME_SIZE)
-        .on(file)
-        .withValue(file.filename().length())
-        .save();
+    if (context.config().getBoolean(HelloWorldProperties.HELLO_KEY).orElse(false)) {
+      // print log only if property is set to true
+      Loggers.get(getClass()).info("Hello World!");
     }
   }
+
 }
