+++ src/main/java/org/sonarsource/plugins/bml/web/MyPluginPageDefinition.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.plugins.example.web;
+package org.sonarsource.plugins.bml.web;
 
 import org.sonar.api.web.page.Context;
 import org.sonar.api.web.page.Page;
